Add path dependencies to workspace-local crates.
These will be used in the workspace, but on crates.io the correct versions will be used.

Cargo workspaces hopefully handles the dependencies correctly.

This should fix the problem of having multiple variants of the same crate with the same version (current local, and current crates.io) when building docs.