Wrap petgraph into own type instead of erasing it.
This way, we can implement traits on that type that petgraph might not implement.
Additionally, before we only were able to have the erased type implement traits that are possible to implement on a very generic graph, since e.g. Eq is restricting the data associated with nodes and edges. If we wanted to restrict the former constructor to Eq, we would restrict it to graphs with comparable node and edge data, which is too restrictive in general.